refactor: View로 분리했던 UIObjects -> UIViewController로 이동
